Display the total marks scored in 3 subjects
Calculate the total marks scored in 3 subjects (100 marks each) and display the score out of 300 marks.

For example:
  english = 56
  maths = 73
  science = 79
  Total Marks = 208

Follow the steps given below to achieve the desired result:
Step 1: Declare three variables english, maths and science respectively and store any numerical values in it.
Step 2: Calculate the total marks by adding the values stored in the variables using the + operator. Store the result in the variable total.
Step 3: Print the total score.