Natural Numbers from 2011 - 2021
Here you need to print all the natural numbers from 2011 - 2021 (including both 2011 and 2021) using a while loop.

Follow the steps given below to achieve the desired result:

Step 1: Store the starting number i.e. 2011 in a variable num.

Step 2: Define the condition for while loop. This loop must be executed till the value stored in variable num is less than or equal to 2021. So the condition for while loop will be while(num)<=2021.

Step 3: Inside the while loop, first print the value stored in num variable, then increment the value of num by 1 to print the next consecutive number.