Football Points
Get the number of wins of a football team as an input. Write a program to calculate the total number of win points the football team has obtained so far.

Every single win gains 3 points to the team.

Example:
Sample Input:
            Enter the number of matches won by the football team: 5
Sample Output:
            The total number of points for the team is 15

Note: The value of win should be greater than or equal to 0.

Follow the steps given below to achieve the desired result:

Step 1: Ask the user to enter the number of matches won by the football team using input() function and convert it into integer using int() function. Store the user input in a variable wins as follows:

wins=int(input(" Enter the number of matches won by the football team:"))

Step 2: Calculate the points gained on winning the matches by multiplying the number of wins (stored in wins variable) by 3 using the * operator (as each win gives 3 points to the team). Store the result in a variable (lets say win_points)

Step 3: Print the value stored in the variable win_points.