Increment a Number by 50
Write a function to increment a number by 50.

For example: if the number entered is 12 then expected output is: 12 + 50 = 62.

Follow the steps given below to achieve the desired result:

Step 1: Ask the user to enter a number. Convert the data type of user input to integer and then store it in a variable num.

Step 2: Create a function (using the def keyword) and name it increment() which takes one input, n.

Step 3: Inside the function, add the input parameter n with 50 using the + operator and store the result in a variable result.

Step 4: Print the result in the output using the return keyword. If you don't write the return statement, then this function will only add the numbers but won't give any output.

Step 5: Call the increment() function (outside the function) and pass the variables num (the variable which holds the user input) as an input to this function.