Multiplication of Two Numbers
Write a function to print the multiplication of two numbers.

For example, if num_1 = 5 and num2 = 15, then result = 5 * 15 = 75

Follow the steps given below to achieve the desired result:

Step 1: Ask the user to enter the first number. Convert the data type of user input to integer using the int() function and then store it in a variable num_1.

Step 2: Ask the user to enter the second number. Convert the data type of user input to integer using the int() function and then store it in a variable num_2.

Step 3: Create a function (using the def keyword) and name it multiplier() which takes two inputs, num1 and num2.

Step 4: Inside the function, multiply the input parameters num1 and num2 using the * operator and store the result in a variable result.

Step 5: Print the result in the output using the return keyword. If you don't write the return statement, then this function will only add the numbers but won't give any output.

Step 6: Now call the multiplier() function (outside the function) and pass the variables num_1 and num_2 (the variable which holds the user input) as an input to this function.