Even or Odd
Accept a number from the user and check whether it is even or odd.

For Example:

Input: 24
Ouput: Even number

Follow the steps given below to achieve the desired result:

Step 1: Ask the user to enter a number.

Step 2: Convert the data type of user input to integer and then store it in a variable num.

Step 3: Use if statement to determine whether the entered number is odd or even. For this, use modulo division (%). The result of a modulo division is the remainder of an integer division. If the number is completely divisible by 2, then print Even number else print Odd number.