Multiplication table
Take input from the user and display the multiplication table.

For example: if the number entered is 12 then expected output is:

12 x 1 = 12
12 x 2 = 24
....
12 x 10 = 120

Follow the steps given below to achieve the desired result:

Step 1: Ask the user to enter a number. Convert the data type of user input to integer and then store it in a variable num.

Step 2: Initialize a variable i=1 (This variable is to make sure that the while loop is executed only upto 10, starting from 1).

Step 3: Create a while loop. This loop will be executed 10 times. So, the condition for while loop will be i < 11.

Step 4: Multiply every value of the i variable with the value of the num variable and store in the multiply variable.

Step 5: Print the result stored in the multiply variable using the print() function and increment the value of i by 1 inside the while loop. The execution stops once the condition for while loop becomes invalid.