Print the series
Write a function to print the following series of a number:

Series of numbers: 0, 1, 1, 2, 3, 5, 8, 13, 21, 34, ...

In this series:
The next number is found by adding the two numbers before it.

For example: Number '2' is obtained by adding the two numbers before it i.e.1+1. Similarly, Number '3' is obtained by adding the two numbers before it i.e.1+2. And Number '5' is obtained after adding 2+3, and so on.

Example:
series of 3: 0, 1, 1. Here 3 is considered as the length of the series.
series of 7: 0, 1, 1, 2, 3, 5, 8. Here 7 is considered as the length of the series.

Steps to obtain the above series:
Step 1: Create a function series() and pass n as a parameter. Here n represents the length of the series.
Step 2: Initialize the first two values of the series in the variables n1 and n2.
Step 3: Initialize the count variable with a value of 0 which gets incremented further in the while loop till the length of the series.
Step 4: If n is less than or equal to 0 then prompt a message saying Please enter a positive integer.
Step 5: If the number n is greater than 0, run a while loop such that the loop iterates until the condition count < n is True.
Step 6: Inside this while loop, perform the following tasks:
     Print the value stored in variable n1.
     Create a temporary variable nth. Store the addition of n1 and n2 in nth i.e. nth=n1+n2.
     Assign the value of n2 to n1 and nth to n2.
     Update the value of count variable by incrementing it by 1.
Step 7: Call the series() function and pass a number as the parameter. This number indicates the length of the series.