Find the missing number
Create a function that takes a list of numbers between 1 and 10 (excluding one number) and returns the missing number.

Notes:
The list of numbers will be unsorted (not in order).
Only one number will be missing.

Examples
missing_num([1, 2, 3, 4, 6, 7, 8, 9, 10]) ➞ 5
missing_num([7, 2, 3, 6, 5, 9, 1, 4, 8]) ➞ 10
missing_num([10, 5, 1, 2, 4, 6, 8, 3, 9]) ➞ 7

Steps to solve the above problem
Step 1: Create a function missing_num() and pass lst as a parameter inside this function, where lst is the list of numbers from 1 - 10 with one missing number.
Step 2: Run a for loop in range 1 - 11 i.e. from 1 to 10. Inside the for loop, check whether the number n is present in the list lst using an if condition.
Step 3: Inside the for loop, check whether each number is present in list lst using an if condition and in and not in operator. If the number is not present in the list lst[], then print that number in the output.