Multiplying numbers of the list
Create a list which contains some numbers and multiply all the numbers in the list.

Constraint: The list should not be empty.

For Example:
list[1,2,3] will give '6'
list[2,3,7] will give '42'

Follow the steps given below to achieve the desired result:

Step 1: Declare a list num_list containing some numbers.
Step 2: Use len() function to find the length of the list and store it in length variable.
Step 3: Initialize two variables:
             i = 0 (To hold index number needed for list indexing. As index always starts with 0, that is why this variable is initialized to 0.)
            result = 1 (To store the product of numbers)
Step 4: Use a while loop. This loop continues to iterate till all the numbers of the list are multiplied i.e. the condition for while loop to iterate is i<length.
Step 5: Inside this while loop, perform the following tasks:
             Multiply the value stored in result variable with each element of the list. To access each number of the list, use list indexing.
             Increment the value of variable i by 1.
Step 6: Print the result.